+++ sdk/identity/azure-identity/CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Bugs Fixed
 
+- Fixed the request sent in `AzurePipelinesCredential` so it doesn't result in a redirect response when an invalid system access token is provided.
+
 ### Other Changes
 
 - Allow certain response headers to be logged in `AzurePipelinesCredential` for diagnostics and include them in the exception message.